Userauth

+++ controllers/nominations.js
@@ -7,8 +7,6 @@ const nominationSeed = require('../models/seed_nominations')
 const Screening = require('../models/screening')
 const Movie = require('../models/movies')
 
-const maintenance = require('../models/maintenance') 
-
 const genres = require('../models/genres')
 
 const nominators = require('../models/nominators')
@@ -50,9 +48,6 @@ router.get('/:id/json', (req, res) => {
 
 //INDEX
 router.get('/', (req, res) => {
-  if (maintenance) {
-    res.render('maintenance.ejs', {tabTitle: 'FFS Maintenance Mode'})
-} else {
     Screening.find({}, (err, screenings) => {
       if(err) console.log('screenings error' + err);
       Movie.find({}, (err, movies) => {
@@ -75,6 +70,8 @@ router.get('/', (req, res) => {
             }
           })
           res.render('nominations/index.ejs', {
+            user: req.user,
+            sessionID: req.sessionID,
             tabTitle: 'FFS Nominations',
             nominations: nominations,
             screenings: screenings,
@@ -83,20 +80,25 @@ router.get('/', (req, res) => {
         })
       })
     })
-  }
 })
 
 //NEW
 router.get('/new', (req, res) => {
-  if (maintenance) {
-    res.render('maintenance.ejs', {tabTitle: 'FFS Maintenance Mode'})
-} else {
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+    res.render('no_access.ejs', {
+        user: req.user,
+        sessionID: req.sessionID,
+        tabTitle: 'Not Authorized'
+    })
+  } else {
     Screening.find({}, (err, screenings) => { //show screenings in reverse date order
       screenings.sort((a,b) => {
         if (a.date > b.date) { return 1 } else { return -1 };
       })
       Movie.find({}, (err, movies) => {
         res.render('nominations/create_nomination.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
           tabTitle: 'Create Nomination',
           screenings: screenings,
           movies: movies,
@@ -194,126 +196,147 @@ const createNominationFromDB = (nomObj, movie, screening, res) => { //create a n
 
 //CREATE
 router.post('/', (req, res) => {
-  // console.log(req.body)
-  let nomObj = {} // build a properly formatted nomination prior to calling Nomination.create
-  // all Nominations conform to model with keys that include screening, nominee, nominator, blurb, and winner.
-  nomObj.nominator = req.body.nominator
-  nomObj.blurb = req.body.blurb
-  nomObj.winner = req.body.winner === "on" ? true : false
-
-  if (req.body.nominee) { // if user chooses an unwatched movie from the database to submit
-    console.log('User chose an unwatched movie from existing db of movies. Creating nomination...');
-    Screening.findOne({weekID: req.body.screeningID}, (err, foundScreening) => { //assign the screeningID to the new nomination object
-      nomObj.screening = foundScreening._id
-      Movie.findOne({title: req.body.nominee}, (err, foundMovie) => { // find the movie in the database and assign it to the nomination object
-        nomObj.nominee = foundMovie._id
-        // Create the new nomination
-        createNominationFromDB(nomObj, foundMovie, foundScreening, res);
-      }) 
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+    res.render('no_access.ejs', {
+        user: req.user,
+        sessionID: req.sessionID,
+        tabTitle: 'Not Authorized'
     })
-  } else { //if it's a first-time nominee
-    console.log('taking the else road')
-    //check that it's not a duplicate
-    Movie.find({}, (err, allMovies) => {
-      for (let movie of allMovies) {
-          if (movie.title === req.body.title) {
-              res.send("<h1>Cannot create movie. Movie's title already exists in database. Check existing movie list.<h1>")
-              return
-          }
-      }
-      let movieObj = req.body
-      let castArray = req.body.cast.split(', ')
-      movieObj.cast = castArray
-      movieObj.year = +req.body.year // convert to number with unary operator (+)
-      movieObj.origNominator = req.body.nominator
-      movieObj.allNominators = [req.body.nominator]
-      movieObj.nominations = []
-      
-      Screening.findOne({weekID: req.body.screeningID}, (err, foundScreening) => {
-        if (req.body.winner === "on") { // if the nomination includes a newly added movie that also was a winner, make sure the movie data reflects that when created
-          movieObj.screened = true
-          movieObj.screening = foundScreening._id
+  } else {
+    // console.log(req.body)
+    let nomObj = {} // build a properly formatted nomination prior to calling Nomination.create
+    // all Nominations conform to model with keys that include screening, nominee, nominator, blurb, and winner.
+    nomObj.nominator = req.body.nominator
+    nomObj.blurb = req.body.blurb
+    nomObj.winner = req.body.winner === "on" ? true : false
+
+    if (req.body.nominee) { // if user chooses an unwatched movie from the database to submit
+      console.log('User chose an unwatched movie from existing db of movies. Creating nomination...');
+      Screening.findOne({weekID: req.body.screeningID}, (err, foundScreening) => { //assign the screeningID to the new nomination object
+        nomObj.screening = foundScreening._id
+        Movie.findOne({title: req.body.nominee}, (err, foundMovie) => { // find the movie in the database and assign it to the nomination object
+          nomObj.nominee = foundMovie._id
+          // Create the new nomination
+          createNominationFromDB(nomObj, foundMovie, foundScreening, res);
+        }) 
+      })
+    } else { //if it's a first-time nominee
+      console.log('taking the else road')
+      //check that it's not a duplicate
+      Movie.find({}, (err, allMovies) => {
+        for (let movie of allMovies) {
+            if (movie.title === req.body.title) {
+                res.send("<h1>Cannot create movie. Movie's title already exists in database. Check existing movie list.<h1>")
+                return
+            }
         }
-        nomObj.screening = foundScreening._id // regardless, the nomination will be associated with the screening date indicated in the form
-        Movie.create(movieObj, (err, createdMovie) => { // finally, you can create the movie
-          if(err) console.log(err);
-          //assign the newly created movie to the nomination object
-          nomObj.nominee = createdMovie._id
-          Nomination.create(nomObj, (err, createdNomination) => { // and then you can create the nomination
+        let movieObj = req.body
+        let castArray = req.body.cast.split(', ')
+        movieObj.cast = castArray
+        movieObj.year = +req.body.year // convert to number with unary operator (+)
+        movieObj.origNominator = req.body.nominator
+        movieObj.allNominators = [req.body.nominator]
+        movieObj.nominations = []
+        
+        Screening.findOne({weekID: req.body.screeningID}, (err, foundScreening) => {
+          if (req.body.winner === "on") { // if the nomination includes a newly added movie that also was a winner, make sure the movie data reflects that when created
+            movieObj.screened = true
+            movieObj.screening = foundScreening._id
+          }
+          nomObj.screening = foundScreening._id // regardless, the nomination will be associated with the screening date indicated in the form
+          Movie.create(movieObj, (err, createdMovie) => { // finally, you can create the movie
             if(err) console.log(err);
-            // now update the movie and the screening to include the newly created info for both the nomination and the movie
-            Movie.findByIdAndUpdate(createdMovie._id, {$set: {nominations: [createdNomination._id]}}, {new: true}, (err, updatedMovie) => { //associate nomination with movie
-              Screening.findByIdAndUpdate(foundScreening._id, {$set: {nominations: [...foundScreening.nominations, createdNomination._id]}}, {new: true}, (err, updatedScreening) => { //associate nomination with screening
-                err ? console.log(err) : console.log(updatedScreening)
-                if(movieObj.screened === true) { //only if movie was a winning nom, set the screening selection to reflect that
-                  Screening.findByIdAndUpdate(foundScreening._id, {$set: {selection: updatedMovie._id}}, {new: true}, (err, updatedScreening2) => {
-                    err ? console.log(err) : console.log(updatedScreening2);
-                    console.log('Set selection for screening: ' + updatedScreening2);
-                  })
-                }
-                console.log("Movie created: " + updatedMovie);
-                console.log("Nomination created: " + createdNomination);
-                res.redirect('/nominations');
+            //assign the newly created movie to the nomination object
+            nomObj.nominee = createdMovie._id
+            Nomination.create(nomObj, (err, createdNomination) => { // and then you can create the nomination
+              if(err) console.log(err);
+              // now update the movie and the screening to include the newly created info for both the nomination and the movie
+              Movie.findByIdAndUpdate(createdMovie._id, {$set: {nominations: [createdNomination._id]}}, {new: true}, (err, updatedMovie) => { //associate nomination with movie
+                Screening.findByIdAndUpdate(foundScreening._id, {$set: {nominations: [...foundScreening.nominations, createdNomination._id]}}, {new: true}, (err, updatedScreening) => { //associate nomination with screening
+                  err ? console.log(err) : console.log(updatedScreening)
+                  if(movieObj.screened === true) { //only if movie was a winning nom, set the screening selection to reflect that
+                    Screening.findByIdAndUpdate(foundScreening._id, {$set: {selection: updatedMovie._id}}, {new: true}, (err, updatedScreening2) => {
+                      err ? console.log(err) : console.log(updatedScreening2);
+                      console.log('Set selection for screening: ' + updatedScreening2);
+                    })
+                  }
+                  console.log("Movie created: " + updatedMovie);
+                  console.log("Nomination created: " + createdNomination);
+                  res.redirect('/nominations');
+                })
               })
             })
           })
         })
       })
-    })
+    }
   }
-    
 })
 
 router.post('/initial-info', (req, res) => {
-  let nominator = req.body.nominator
-  let screeningID = req.body.screening
-  let nomination_type = req.body.nomination_type
-  // console.log(nomination_type)
-  Screening.find({weekID: screeningID}, (err, screening) => {
-    screening = screening[0]
-    Movie.find({}, (err, movies) => {
-      movies.sort((a,b) => {
-        if(a.title > b.title) {
-          return 1
-        } else {
-          return -1
-        }
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+      res.render('no_access.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Not Authorized'
       })
-      res.render('nominations/new.ejs', {
-        tabTitle: 'Continue Nomination',
-        screening: screening,
-        movies: movies,
-        genres: genres,
-        nominator: nominator,
-        nomination_type: nomination_type
+  } else {
+    let nominator = req.body.nominator
+    let screeningID = req.body.screening
+    let nomination_type = req.body.nomination_type
+    // console.log(nomination_type)
+    Screening.find({weekID: screeningID}, (err, screening) => {
+      screening = screening[0]
+      Movie.find({}, (err, movies) => {
+        movies.sort((a,b) => {
+          if(a.title > b.title) {
+            return 1
+          } else {
+            return -1
+          }
+        })
+        res.render('nominations/new.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Continue Nomination',
+          screening: screening,
+          movies: movies,
+          genres: genres,
+          nominator: nominator,
+          nomination_type: nomination_type
+        })
       })
     })
-  })
+  }
 })
 
 //SHOW
 router.get('/:id', (req, res) => {
-  if (maintenance) {
-    res.render('maintenance.ejs', {tabTitle: 'FFS Maintenance Mode'})
-  } else {
     Nomination.findById(req.params.id, (err, foundNomination) => {
       res.render('nominations/show.ejs', {
+        user: req.user,
+        sessionID: req.sessionID,
         nomination: foundNomination,
         tabTitle: foundNomination.nominee.title + ' | Nomination' 
       })
     }).populate("screening").populate("nominee")
-  }
 })
 
 
 //EDIT
 router.get('/:id/edit', (req, res) => {
-  if (maintenance) {
-    res.render('maintenance.ejs', {tabTitle: 'FFS Maintenance Mode'})
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+      res.render('no_access.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Not Authorized'
+      })
   } else {
     Screening.find({}, (err, allScreenings) => { 
       Nomination.findById(req.params.id, (err, foundNom) => {
         res.render('nominations/edit.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
           tabTitle: foundNom.nominee.title + " | Edit Nomination",
           nomination: foundNom,
           screenings: allScreenings,
@@ -326,16 +349,33 @@ router.get('/:id/edit', (req, res) => {
 
 //CONFIRM DELETE
 router.get('/:id/confirm-delete', (req, res) => {
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+      res.render('no_access.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Not Authorized'
+      })
+  } else {
     Nomination.findById(req.params.id, (err, foundNom) => {
       res.render('nominations/confirm_delete.ejs', {
+        user: req.user,
+        sessionID: req.sessionID,
         tabTitle: 'Confirm delete?',
         nomination: foundNom
       })
     }).populate("screening").populate("nominee")
+  }
 })
   
 //DELETE
 router.delete('/:id', (req, res) => {
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+      res.render('no_access.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Not Authorized'
+      })
+  } else {
     Nomination.findByIdAndRemove(req.params.id, (err, deletedNomination) => {
       //Find and remove that nomination from its related screening
       Screening.findOneAndUpdate({weekID: deletedNomination.screening}, {$pull: {nominations: deletedNomination}}, {new: true}, (err, updatedScreening) => {
@@ -396,15 +436,22 @@ router.delete('/:id', (req, res) => {
         }
       })
     })
+  }
 })
 
 //UPDATE
 router.put('/:id', (req, res) => {
-    
+  if (!req.isAuthenticated() || req.user.role === 'visitor') {
+      res.render('no_access.ejs', {
+          user: req.user,
+          sessionID: req.sessionID,
+          tabTitle: 'Not Authorized'
+      })
+  } else {
     console.log(req.body)
     let checked = req.body.winner === "on" ? true : false
     req.body.winner = checked
-
+  
     Nomination.findById(req.params.id).populate("screening").populate("nominee").exec((err, preUpdateNom) => {
       if(err) console.log(err.message);
       console.log(preUpdateNom);
@@ -421,7 +468,7 @@ router.put('/:id', (req, res) => {
             }
           })
           let earliestNom = movieNominations[0]
-
+  
           if (earliestNom.screening.weekID === preUpdateNom.screening.weekID) { 
             earliestNom = movieNominations[1] // if the nomination we are updating the screening date for is the earliest current nomination, we now need to find the next earliest nomination, i.e. movieNominations[1]. Otherwise, we will be including the old screening date in the comparison that follows.
           }
@@ -430,7 +477,7 @@ router.put('/:id', (req, res) => {
           if (earliestNom.screening.weekID < req.body.screening) { // Check whether the updating nom is in fact earliest.
               isEarliestNom = false
           }
-
+  
           if (isEarliestNom) { //if the updating nom is the earliest nom, update the origNominator field in the movie with the nominator provided in the edit form
             Movie.findByIdAndUpdate(preUpdateNom.nominee._id, {$set: {origNominator: req.body.nominator}}, {new: true}, (err, updatedMovie) => {
               console.log('New earliest nomination for movie, origNominator updated: ' + updatedMovie);
@@ -443,7 +490,7 @@ router.put('/:id', (req, res) => {
           // 
         })
       }
-
+  
       
         
         //what to do if updated nom is for different nominator. must splice nominator name into allNominators of the movie, removing oldNominator in the process.
@@ -456,7 +503,7 @@ router.put('/:id', (req, res) => {
             }) 
           })
         }
-
+  
         if(preUpdateNom.winner !== req.body.winner) { //if winner status has changed, update appropriately based on which way it changed
           Screening.findOne({weekID: req.body.screening}, (err, nominatedScreening) => {
             req.body.screening = nominatedScreening._id 
@@ -493,6 +540,8 @@ router.put('/:id', (req, res) => {
           })
         }
     })
+  }
+    
   })
 
 // DELETE APP-BREAKING NOMS BY ID, USE WITH CAUTION
